SWITCHYARD-1069: Separate picketlink from core security classes
