+++ security/base/src/main/java/org/switchyard/security/SecurityContext.java
