+++ security/base/src/main/java/org/switchyard/security/SecurityUtil.java
