+++ security/base/src/main/java/org/switchyard/security/callback/NamePasswordCallbackHandler.java
