+++ security/base/src/main/java/org/switchyard/security/credential/CertificateCredential.java
