+++ security/base/src/main/java/org/switchyard/security/credential/extract/CredentialsExtractor.java
