+++ security/base/src/main/java/org/switchyard/security/principal/Role.java
