+++ security/base/src/main/resources/META-INF/services/org.switchyard.security.SecurityProvider
