+++ security/base/src/test/java/org/switchyard/security/credential/extract/SOAPMessageCredentialsExtractorTests.java
