+++ security/base/src/test/resources/org/switchyard/security/credential/extract/AuthorizationHeaderCredentialsExtractorTests-Digest.txt
