+++ security/base/src/test/resources/org/switchyard/security/credential/extract/SOAPMessageCredentialsExtractorTests-Assertion.xml
