+++ security/jboss/src/main/java/org/switchyard/security/jboss/callback/STSTokenCallbackHandler.java
@@ -16,52 +16,45 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, 
  * MA  02110-1301, USA.
  */
-package org.switchyard.security.callback;
+package org.switchyard.security.jboss.callback;
 
 import java.io.IOException;
 import java.util.Set;
 
 import javax.security.auth.callback.Callback;
-import javax.security.auth.callback.NameCallback;
-import javax.security.auth.callback.PasswordCallback;
 import javax.security.auth.callback.UnsupportedCallbackException;
 
-import org.jboss.security.auth.callback.ObjectCallback;
-import org.switchyard.security.credential.CertificateCredential;
+import org.picketlink.identity.federation.core.wstrust.auth.TokenCallback;
+import org.switchyard.security.callback.SwitchYardCallbackHandler;
+import org.switchyard.security.credential.AssertionCredential;
 import org.switchyard.security.credential.Credential;
 
 /**
- * CertificateCallbackHandler.
+ * STSTokenCallbackHandler.
  *
  * @author David Ward &lt;<a href="mailto:dward@jboss.org">dward@jboss.org</a>&gt; &copy; 2012 Red Hat Inc.
  */
-public class CertificateCallbackHandler extends SwitchYardCallbackHandler {
+public class STSTokenCallbackHandler extends SwitchYardCallbackHandler {
 
     /**
-     * Constructs a new CertificateCallbackHandler.
+     * Constructs a new STSTokenCallbackHandler.
      */
-    public CertificateCallbackHandler() {}
+    public STSTokenCallbackHandler() {}
 
     /**
      * {@inheritDoc}
      */
     @Override
     public void handle(Callback[] callbacks) throws IOException, UnsupportedCallbackException {
-        String alias = getProperty("alias", true);
-        String keyPassword = getProperty("keyPassword", true);
         Set<Credential> credentials = getCredentials();
         if (credentials == null) {
             throw new IllegalStateException("Credentials not set");
         }
         for (Callback cb : callbacks) {
-            if (cb instanceof NameCallback) {
-                ((NameCallback)cb).setName(alias);
-            } else if (cb instanceof PasswordCallback) {
-                ((PasswordCallback)cb).setPassword(keyPassword.toCharArray());
-            } else if (cb instanceof ObjectCallback) {
+            if (cb instanceof TokenCallback) {
                 for (Credential cred : credentials) {
-                    if (cred instanceof CertificateCredential) {
-                        ((ObjectCallback)cb).setCredential(((CertificateCredential)cred).getCertificate());
+                    if (cred instanceof AssertionCredential) {
+                        ((TokenCallback)cb).setToken(((AssertionCredential)cred).getAssertion());
                     }
                 }
             }
